Separate secrets file into encoding utility and authorize file
