+++ modules/utilityModules/cookie.js
@@ -4,7 +4,7 @@
 const path = require("path"); // URI and local file paths
 
 // Utility Modules
-const utilityModulesPath = path.join(__dirname);
+const utilityModulesPath = __dirname;
 const logger = require(path.join(utilityModulesPath, "logger.js"));
 const environment = require(path.join(utilityModulesPath, "environment.js"));
